Move various payload helpers out
Move helpers only used by apiserver/payloadshookcontext into that
package, out of payload/api/private.
