+++ payload/api/private/helpers_test.go
@@ -23,50 +23,6 @@ type internalHelpersSuite struct {
 
 var _ = gc.Suite(&internalHelpersSuite{})
 
-func (internalHelpersSuite) TestNewPayloadResultOkay(c *gc.C) {
-	id := "ce5bc2a7-65d8-4800-8199-a7c3356ab309"
-	result := private.NewPayloadResult(id, nil)
-
-	c.Check(result, jc.DeepEquals, params.PayloadResult{
-		Entity: params.Entity{
-			Tag: names.NewPayloadTag(id).String(),
-		},
-		Payload:  nil,
-		NotFound: false,
-		Error:    nil,
-	})
-}
-
-func (internalHelpersSuite) TestNewPayloadResultError(c *gc.C) {
-	id := "ce5bc2a7-65d8-4800-8199-a7c3356ab309"
-	err := errors.New("<failure>")
-	result := private.NewPayloadResult(id, err)
-
-	c.Check(result, jc.DeepEquals, params.PayloadResult{
-		Entity: params.Entity{
-			Tag: names.NewPayloadTag(id).String(),
-		},
-		Payload:  nil,
-		NotFound: false,
-		Error:    common.ServerError(err),
-	})
-}
-
-func (internalHelpersSuite) TestNewPayloadResultNotFound(c *gc.C) {
-	id := "ce5bc2a7-65d8-4800-8199-a7c3356ab309"
-	err := errors.NotFoundf("payload %q", id)
-	result := private.NewPayloadResult(id, err)
-
-	c.Check(result, jc.DeepEquals, params.PayloadResult{
-		Entity: params.Entity{
-			Tag: names.NewPayloadTag(id).String(),
-		},
-		Payload:  nil,
-		NotFound: true,
-		Error:    common.ServerError(err),
-	})
-}
-
 func (internalHelpersSuite) TestAPI2ResultOkay(c *gc.C) {
 	id := "ce5bc2a7-65d8-4800-8199-a7c3356ab309"
 	result, err := private.API2Result(params.PayloadResult{
@@ -167,101 +123,5 @@ func (internalHelpersSuite) TestAPI2ResultNotFound(c *gc.C) {
 		Payload:  nil,
 		NotFound: false,
 		Error:    result.Error, // The actual error is checked above.
-	})
-}
-
-func (internalHelpersSuite) TestResult2apiOkay(c *gc.C) {
-	id := "ce5bc2a7-65d8-4800-8199-a7c3356ab309"
-	result := private.Result2api(payload.Result{
-		ID:       id,
-		Payload:  nil,
-		NotFound: false,
-		Error:    nil,
-	})
-
-	c.Check(result, jc.DeepEquals, params.PayloadResult{
-		Entity: params.Entity{
-			Tag: names.NewPayloadTag(id).String(),
-		},
-		Payload:  nil,
-		NotFound: false,
-		Error:    nil,
-	})
-}
-
-func (internalHelpersSuite) TestResult2apiInfo(c *gc.C) {
-	id := "ce5bc2a7-65d8-4800-8199-a7c3356ab309"
-	result := private.Result2api(payload.Result{
-		ID:       id,
-		NotFound: false,
-		Error:    nil,
-		Payload: &payload.FullPayloadInfo{
-			Payload: payload.Payload{
-				PayloadClass: charm.PayloadClass{
-					Name: "foobar",
-					Type: "type",
-				},
-				ID:     "idfoo",
-				Status: payload.StateRunning,
-				Unit:   "a-application/0",
-			},
-			Machine: "1",
-		},
-	})
-
-	c.Check(result, jc.DeepEquals, params.PayloadResult{
-		Entity: params.Entity{
-			Tag: names.NewPayloadTag(id).String(),
-		},
-		NotFound: false,
-		Error:    nil,
-		Payload: &params.Payload{
-			Class:   "foobar",
-			Type:    "type",
-			ID:      "idfoo",
-			Status:  payload.StateRunning,
-			Unit:    "unit-a-application-0",
-			Machine: "machine-1",
-		},
-	})
-}
-
-func (internalHelpersSuite) TestResult2apiError(c *gc.C) {
-	id := "ce5bc2a7-65d8-4800-8199-a7c3356ab309"
-	err := errors.New("<failure>")
-	result := private.Result2api(payload.Result{
-		ID:       id,
-		Payload:  nil,
-		NotFound: false,
-		Error:    err,
-	})
-
-	c.Check(result, jc.DeepEquals, params.PayloadResult{
-		Entity: params.Entity{
-			Tag: names.NewPayloadTag(id).String(),
-		},
-		Payload:  nil,
-		NotFound: false,
-		Error:    common.ServerError(err),
-	})
-}
-
-func (internalHelpersSuite) TestResult2apiNotFound(c *gc.C) {
-	id := "ce5bc2a7-65d8-4800-8199-a7c3356ab309"
-	err := errors.NotFoundf("payload %q", id)
-	result := private.Result2api(payload.Result{
-		ID:       id,
-		Payload:  nil,
-		NotFound: false,
-		Error:    err,
-	})
-
-	c.Check(result, jc.DeepEquals, params.PayloadResult{
-		Entity: params.Entity{
-			Tag: names.NewPayloadTag(id).String(),
-		},
-		Payload:  nil,
-		NotFound: false,
-		Error:    common.ServerError(err),
 	})
 }
